// submodules/common-dto/api-client/main/models/MainSubjectPostRequestDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MainSubjectPostRequestDto = {
    /**
     * Slug
     */
    slug: string;
    /**
     * Localized title of subject
     */
    title: string;
    /**
     * ID of image
     */
    image_id: string | null;
};