// submodules/common-dto/api-client/main/models/MainMaterialPostRequestDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MainMaterialPostRequestDto = {
    /**
     * Slug
     */
    slug: string;
    /**
     * Localized title of material
     */
    title: string;
    /**
     * ID of image
     */
    image_id: string | null;
};