// submodules/common-dto/api-client/main/models/MainKeywordsPostRequestDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MainKeywordsPostRequestDto = {
    /**
     * Localized title of keyword
     */
    title: string;
    /**
     * ID of image
     */
    image_id: string | null;
};