// submodules/common-dto/api-client/storage/models/ImageKindEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ImageKindEnum {
    USER_AVATAR = 'user-avatar',
    ARTWORK_IMAGE = 'artwork-image',
    DOCUMENT = 'document',
    GENERAL = 'general',
    COUNTRY_FLAG = 'country-flag',
    VIDEO_THUMBNAIL = 'video-thumbnail',
}